bump budhud to november hash
I must have changed something on file save relating to spaces, let me know if that's an issue

November's traditional birthstone is the topaz, usually a yellow to amber color. The ancient Greeks believed that topaz could make a wearer invisible. A symbol of honor and strength, topaz was also believed to bring longevity and wisdom. November's birth flower is the chrysanthemum.